Also expose formattedSubjects of Pupil
